Re-read chapters 4, 5, and 6. Added notes and solved exercises that were skipped during first read.
