--- Introduction_to_Programming_in_Go/ch05_control-structures/Fah2Cel/main.go
@@ -1,15 +1,0 @@
-package main
-
-import "fmt"
-
-func main() {
-  //: Ask user for Fahrenheit measurement
-  fmt.Println("Enter Fahrenheit: ")
-  var fahrenheit float64
-  fmt.Scanf("%f", &fahrenheit)
-
-  celsius := (fahrenheit - 32) * 5 / 9
-
-  fmt.Println(celsius)
-
-}
