--- Introduction_to_Programming_in_Go/ch05_control-structures/feet2meters/main.go
@@ -1,13 +1,0 @@
-package main
-
-import "fmt"
-
-func main() {
-  fmt.Print("Enter feet: ")
-  var feet float64
-  fmt.Scanf("%f", &feet)
-
-  meters := feet * 0.3048
-
-  fmt.Println(meters)
-}
